Super admin

+++ src/components/UserList.js
@@ -1,7 +1,13 @@
 import React, { useState, useEffect } from "react";
-import { FlatList, Text, TouchableOpacity, View, StyleSheet, Alert, TextInput} from "react-native";
-import Swipeable from 'react-native-gesture-handler/Swipeable';
-import Constants from 'expo-constants';
+import {
+  FlatList,
+  Text,
+  TouchableOpacity,
+  View,
+  StyleSheet,
+} from "react-native";
+import Swipeable from "react-native-gesture-handler/Swipeable";
+import Constants from "expo-constants";
 
 import {
   getFirestore,
@@ -10,12 +16,22 @@ import {
   doc,
   updateDoc,
   deleteDoc,
+  getDoc,
+  query,
+  where,
 } from "firebase/firestore";
+import { getAuth, onAuthStateChanged } from "firebase/auth";
+
+const auth = getAuth();
 
 const UserList = () => {
   const [users, setUsers] = useState([]);
   const [loading, setLoading] = useState(true);
+  const [isSuperToggle, setIsSuperToggle] = useState(false);
+  const [currentUser, setCurrentUser] = useState(null);
   const db = getFirestore();
+  const auth = getAuth();
+  const user = auth.currentUser;
 
   const [newSearch, setNewSearch] = useState("");
 
@@ -24,6 +40,48 @@ const UserList = () => {
 
   //here we fatch the user list so we can display them later
   useEffect(() => {
+    const unsubscribe = onAuthStateChanged(auth, (user) => {
+      setCurrentUser(user);
+    });
+    return unsubscribe;
+  }, []);
+
+  useEffect(() => {
+    const fetchData = async () => {
+      const userQuery = query(
+        collection(db, "users"),
+        where("user_id", "==", user.uid)
+      );
+
+      const querySnapshot = await getDocs(userQuery);
+      querySnapshot.forEach((doc) => {
+        if (doc.data().is_super_admin == true) {
+          setIsSuperToggle(!isSuperToggle);
+          console.log(isSuperToggle);
+
+          const fetchUsers = async () => {
+            const querySnapshot = await getDocs(collection(db, "users"));
+            const fetchedUsers = [];
+            querySnapshot.forEach((doc) => {
+              const user = doc.data();
+              user.id = doc.id;
+              fetchedUsers.push(user);
+            });
+
+            setUsers(fetchedUsers);
+            setLoading(false);
+          };
+
+          fetchUsers();
+        } else {
+          setIsSuperToggle(isSuperToggle);
+
+          console.log(isSuperToggle);
+        }
+      });
+    };
+    fetchData().catch((error) => console.log(error));
+
     const fetchUsers = async () => {
       const querySnapshot = await getDocs(collection(db, 'users'));
       const fetchedUsers = [];
@@ -31,7 +89,9 @@ const UserList = () => {
       querySnapshot.forEach((doc) => {
         const user = doc.data();
         user.id = doc.id;
-        fetchedUsers.push(user);
+        if (!user.is_super_admin == true && !user.is_admin == true) {
+          fetchedUsers.push(user);
+        }
       });
 
       setUsers(fetchedUsers);
@@ -54,9 +114,16 @@ const UserList = () => {
   // This edits the admin status of the user. 
   const makeAdmin = async (userId, is_admin) => {
     try {
+      // Retrieve the user's information
       const userRef = doc(db, "users", userId);
+      // Check if the current user is a super admin
+
+      // Update the user's is_admin property in the Firestore database
       await updateDoc(userRef, { is_admin: !is_admin });
+
       console.log("User is now an admin");
+
+      // Update the state or UI with the updated user data
       const updatedUsers = users.map((user) => {
         if (user.id === userId) {
           return { ...user, is_admin: !is_admin };
@@ -84,36 +151,29 @@ const UserList = () => {
   const deleteUser = async (userId) => {
     try {
       const userRef = doc(db, "users", userId);
-      await deleteDoc(userRef);
-      console.log("User deleted successfully");
-      const updatedUsers = users.filter((user) => user.id !== userId);
-      setUsers(updatedUsers);
+      const userDoc = await getDoc(userRef);
+      const userData = userDoc.data();
+      if (
+        currentUser &&
+        (currentUser.uid === userData.created_by ||
+          userData.is_super_admin === true) &&
+        userData.is_super_admin !== true
+      ) {
+        await deleteDoc(userRef);
+        console.log("User deleted successfully");
+        const updatedUsers = users.filter((user) => user.id !== userId);
+        setUsers(updatedUsers);
+      } else {
+        console.log("You are not authorized to perform this action");
+      }
     } catch (error) {
       console.error("Error deleting user:", error);
     }
   };
 
-  const fetchSearchUsers = async (newSearch) => {
-    const querySnapshot = await getDocs(collection(db, 'users'));
-    const fetchedUsers = [];
-
-    querySnapshot.forEach((doc) => {
-      const user = doc.data();
-      if (user.email.includes(newSearch)) {
-        user.id = doc.id;
-        fetchedUsers.push(user);
-      } else {
-        
-      }
-    });
-
-    setUsers(fetchedUsers);
-    setLoading(false);
-  };
-  
   const renderUser = ({ item, index }, onClick) => {
     const closeRow = (index) => {
-      console.log('closerow');
+      console.log("closerow");
       if (prevOpenedRow && prevOpenedRow !== row[index]) {
         prevOpenedRow.close();
       }
@@ -122,88 +182,152 @@ const UserList = () => {
 
     //renders when swiping the user name.
     const renderRightActions = (progress, dragX, onClick) => {
-      return (
-        <View
-          style={{
-            flexDirection: "row",
-            margin: 0,
-            alignContent: 'center',
-            justifyContent: 'center',
-            width: 200,
-          }}>
-            <TouchableOpacity
+      if (isSuperToggle === true) {
+        return (
+          <View
             style={{
-              backgroundColor: item.is_admin ? "gray" : "#3F51B5",
-              padding: 10,
-              borderRadius: 8,
-              marginRight: 8,
-              height: 40,
-              width:125,
+              flexDirection: "row",
+              margin: 0,
               alignContent: "center",
+              justifyContent: "center",
+              width: 200,
             }}
-            onPress={() => createTwoButtonAlertAdmin(item.id, item.is_admin)}
           >
-            <Text style={{ color: "white", fontWeight: "bold" , textAlign:"center"}}>
-              {item.is_admin ? "Remove Admin" : "Make Admin"}
-            </Text>
-          </TouchableOpacity>
-          <TouchableOpacity style={{backgroundColor: "#f44336",padding: 10,borderRadius: 8, height: 40,}} onPress={onClick}><Text style={{ color: "white", fontWeight: "bold" }}>Delete</Text></TouchableOpacity>
-        </View>
-      );
+            <TouchableOpacity
+              style={{
+                backgroundColor: item.is_admin ? "gray" : "#3F51B5",
+                padding: 10,
+                borderRadius: 8,
+                marginRight: 8,
+                height: 40,
+                width: 125,
+                alignContent: "center",
+              }}
+              onPress={() => makeAdmin(item.id, item.is_admin)}
+            >
+              <Text
+                style={{
+                  color: "white",
+                  fontWeight: "bold",
+                  textAlign: "center",
+                }}
+              >
+                {item.is_admin ? "Remove Admin" : "Make Admin"}
+              </Text>
+            </TouchableOpacity>
+            <TouchableOpacity
+              style={{
+                backgroundColor: "#f44336",
+                padding: 10,
+                borderRadius: 8,
+                height: 40,
+              }}
+              onPress={onClick}
+            >
+              <Text style={{ color: "white", fontWeight: "bold" }}>Delete</Text>
+            </TouchableOpacity>
+          </View>
+        );
+      } else {
+        return (
+          <View
+            style={{
+              flexDirection: "row",
+              margin: 0,
+              alignContent: "center",
+              justifyContent: "center",
+              width: 80,
+            }}
+          >
+            <TouchableOpacity
+              style={{
+                backgroundColor: "#f44336",
+                padding: 10,
+                borderRadius: 8,
+                height: 40,
+              }}
+              onPress={onClick}
+            >
+              <Text style={{ color: "white", fontWeight: "bold" }}>Delete</Text>
+            </TouchableOpacity>
+          </View>
+        );
+      }
     };
 
-    return(
+    return (
       <Swipeable
         renderRightActions={(progress, dragX) =>
           renderRightActions(progress, dragX, onClick)
         }
         onSwipeableOpen={() => closeRow(index)}
         ref={(ref) => (row[index] = ref)}
-        rightOpenValue={-100}>
-    <View
-      style={{
-        flexDirection: "row",
-        alignItems: "center",
-        alignSelf: "center",
-        alignContent: "center",
-        justifyContent: 'center',
-        //marginVertical: 20,
-        backgroundColor: "#FEFEFE",
-        height: 70,
-      }}>
-      <Text style={{ flex: 1, fontWeight: 'bold',fontSize: 20, left: 30, color: "black"}}>{item.email}</Text>
-      <Text style={{color: "black", alignSelf: "center",position: "absolute",bottom: -15, height: 40,}}>⎯⎯⎯⎯⎯⎯⎯⎯⎯⎯⎯⎯⎯⎯⎯⎯⎯⎯⎯⎯</Text>
-      <Text style={{color: "black", fontWeight: 'bold',fontSize: 20, alignSelf: "center",position: "absolute", right: 30,}}>{"<"}</Text>
-    </View>
-    </Swipeable>
+        rightOpenValue={-100}
+      >
+        <View
+          style={{
+            flexDirection: "row",
+            alignItems: "center",
+            alignSelf: "center",
+            alignContent: "center",
+            justifyContent: "center",
+            //marginVertical: 20,
+            backgroundColor: "#FEFEFE",
+            height: 70,
+          }}
+        >
+          <Text
+            style={{
+              flex: 1,
+              fontWeight: "bold",
+              fontSize: 20,
+              left: 30,
+              color: "black",
+            }}
+          >
+            {item.email}
+          </Text>
+          <Text
+            style={{
+              color: "black",
+              alignSelf: "center",
+              position: "absolute",
+              bottom: -15,
+              height: 40,
+            }}
+          >
+            ⎯⎯⎯⎯⎯⎯⎯⎯⎯⎯⎯⎯⎯⎯⎯⎯⎯⎯⎯⎯
+          </Text>
+          <Text
+            style={{
+              color: "black",
+              fontWeight: "bold",
+              fontSize: 20,
+              alignSelf: "center",
+              position: "absolute",
+              right: 30,
+            }}
+          >
+            {"<"}
+          </Text>
+        </View>
+      </Swipeable>
     );
   };
 
   return (
-    <View>
-      <View style={{height: 40,}}>
-        <TextInput
-					style={{ backgroundColor: "#C0C0C0",
-          alignSelf: "center",
-          position: "absolute",
-          top: 50,
-          height: 40,
-          marginHorizontal: 50,
-          borderRadius: 10,
-          width: 300,}}
-          placeholder=" Search..."
-					onChangeText={(text) => fetchSearchUsers(text)}
-				/>
-      </View>
+
     <FlatList
       data={users}
-      renderItem={(v) => 
+      renderItem={(v) =>
         renderUser(v, () => {
-          console.log('pressed', v);
-          createTwoButtonAlertDelete(v.item.id)
-        })}
+          console.log("pressed", v);
+          deleteUser(v.item.id);
+        })
+      }
+
       keyExtractor={(item) => item.id}
-      style={{ padding: 16, marginTop: 50}}
+      style={{ padding: 16, marginTop: 50 }}
     />
     </View>
   );
@@ -214,9 +338,9 @@ export default UserList;
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-    justifyContent: 'center',
+    justifyContent: "center",
     paddingTop: Constants.statusBarHeight,
-    backgroundColor: '#ecf0f1',
+    backgroundColor: "#ecf0f1",
     padding: 8,
   },
 });
